obj_accessor without an Object no longer fatals
In the case where we're performing a `get` action on an object attribute within an object, that may not be set, and a default is provided, the default is returned.

This shouldn't have any impact to `set` actions; or, to `get` actions where there is no default; or, to `get` actions where an object is passed.